+++ src/redux/reducers.js
@@ -2,9 +2,11 @@ import { combineReducers } from "redux";
 import { reducer as commonManagerReducer } from "./common/reducer";
 import { reducer as expensesManagerReducer } from "./expensesManager/reducer";
 import { reducer as userManagerReducer } from "./userManager/reducer";
+import { reducer as syncManagerReducer } from "./syncManager/reducer";
 
 export default combineReducers({
   expensesManager: expensesManagerReducer,
   userManager: userManagerReducer,
   commonManager: commonManagerReducer,
+  syncManager: syncManagerReducer,
 });